refactor: include invalidation
