+++ src/api/mutations/account/useEditAccountMutation.ts
@@ -4,7 +4,9 @@ import toast from 'react-hot-toast';
 import httpClient from '@src/api/httpClient';
 import { ENDPOINTS } from '@src/config/endpoints.config';
 import { QUERY_KEYS } from '@src/config/react-query.config';
+import { ProfileActions } from '@src/types/i18n.types';
 import { Profile } from '@src/types/profile.types';
+import { t } from 'i18next';
 
 function useEditAccountMutation(profileId: string) {
   const queryClient = useQueryClient();
@@ -15,7 +17,7 @@ function useEditAccountMutation(profileId: string) {
     },
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.query.user] });
-      toast.success('Updated successfully!');
+      toast.success(t(ProfileActions.profile_change));
     },
   });
 }
