+++ src/api/mutations/wishlist/useDeleteFromWishlistMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 
 import httpClient from '@src/api/httpClient';
 import { ENDPOINTS } from '@src/config/endpoints.config';
-import { QUERY_KEYS } from '@src/config/react-query.config';
+import { QUERY_KEYS, queryClient } from '@src/config/react-query.config';
 import { CommonResponse } from '@src/types/auth.types';
 
 function useDeleteFromWishlistMutation() {
@@ -14,6 +14,15 @@ function useDeleteFromWishlistMutation() {
         ENDPOINTS.wishlist.deleteFromWishlist(accommodationId)
       );
       return data;
+    },
+
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({
+        queryKey: [QUERY_KEYS.query.wishlist],
+      });
+      await queryClient.invalidateQueries({
+        queryKey: [QUERY_KEYS.query.accommodations],
+      });
     },
   });
 }
