+++ src/components/shared/CustomImage.tsx
@@ -2,7 +2,7 @@ import { Box } from '@mui/material';
 import { useEffect, useState } from 'react';
 
 import LoadingImage from '@src/assets/loader.png';
-import ErrorImage from '@src/assets/no-image.png';
+import ErrorImage from '@src/assets/no-image.jpg';
 
 const imageStates = {
   Loading: LoadingImage,
